Implement edit functionality
Implemented part of edit permission functionality with topics.

+++ client/agora/public/js/state/stateManager.js
@@ -2,7 +2,7 @@
 import { addTopicEvent } from "../editorMain.js";
 import { createNewResource, saveResource } from "../controllers/clientResourceController.js";
 import { createNewTopic, saveTopic, getTopic, getResourcesForTopic, deleteTopic } from "../controllers/clientTopicController.js";
-import { createNewWorkspace, saveWorkspace, getWorkspace } from "../controllers/clientWorkspaceController.js";
+import { createNewWorkspace, saveWorkspace, getWorkspace, getPermission } from "../controllers/clientWorkspaceController.js";
 
 /**
  * Client side debugging flags
@@ -19,6 +19,7 @@ let workspace = createNewResource;
 
 // Active topic for the chosen tab
 let activeTopic = createNewTopic();
+
 /*--------------------------------------------------------------------------------*/
 
 /**
@@ -97,8 +98,7 @@ const setActiveTopicAndResources = async function ( topicId ) {
 
 const addNewTopic = async function ( topicName ) {
     ( debug ) ? console.log( "addNewTopic() : Start - topicName: " + topicName ) : null;
-
-    if( getCurrentWorkspace() ) {
+    if( getCurrentWorkspace() && await getPermission( getCurrentWorkspace().workspaceId ) ) {
         // create a new topic
         let newTopic = createNewTopic();
         newTopic.topicName = topicName;
@@ -116,7 +116,7 @@ const addNewTopic = async function ( topicName ) {
         // save the current workspace
         await saveWorkspace( getCurrentWorkspace() );
 
-        
+            
         ( debug ) ? console.log( "addNewTopic() : Complete" ) : null;
         return newTopic;
     }
@@ -129,18 +129,20 @@ const addNewTopic = async function ( topicName ) {
 };
 
 const updateTopicName = async function ( topicId, topicName ) {
-    ( debug ) ? console.log( "saveActiveTopic() : Start" ) : null;
+    if ( await getPermission( getCurrentWorkspace().workspaceId ) ){
+        ( debug ) ? console.log( "saveActiveTopic() : Start" ) : null;
 
-    // get the topic from the current workspace
-    let topic = getCurrentWorkspace().topics.find( topic => topic.topicId === topicId );
+        // get the topic from the current workspace
+        let topic = getCurrentWorkspace().topics.find( topic => topic.topicId === topicId );
 
-    // update the topic name
-    topic.topicName = topicName;
+        // update the topic name
+        topic.topicName = topicName;
 
-    // save the topic
-    await saveTopic( topic );
+        // save the topic
+        await saveTopic( topic );
 
-    ( debug ) ? console.log( "saveActiveTopic() : Start" ) : null;
+        ( debug ) ? console.log( "saveActiveTopic() : Start" ) : null;
+    }
 };
 
 const saveActiveTopic = async function ( ) {
@@ -157,41 +159,44 @@ const saveActiveTopic = async function ( ) {
 };
 
 const deleteTopicFromWorkspace = async function ( topicId ) {
-    ( debug ) ? console.log( "deleteTopicFromWorkspace() : Start for topicId - " + topicId ) : null;
+    if( await getPermission( getCurrentWorkspace().workspaceId ) ){
+        ( debug ) ? console.log( "deleteTopicFromWorkspace() : Start for topicId - " + topicId ) : null;
 
-    // delete the topic and resources
-    await deleteTopic( topicId );
+        // delete the topic and resources
+        await deleteTopic( topicId );
 
-    // remove the topic from the current workspace
-    getCurrentWorkspace().topics = getCurrentWorkspace().topics.filter( topic => topic.topicId !== topicId );
+        // remove the topic from the current workspace
+        getCurrentWorkspace().topics = getCurrentWorkspace().topics.filter( topic => topic.topicId !== topicId );
 
-    // save the current workspace (deletes association with workspace)
-    await saveWorkspace( getCurrentWorkspace() );
+        // save the current workspace (deletes association with workspace)
+        await saveWorkspace( getCurrentWorkspace() );
 
     
     
-    ( debug ) ? console.log( "deleteTopicFromWorkspace() : Complete" ) : null;
+        ( debug ) ? console.log( "deleteTopicFromWorkspace() : Complete" ) : null;
+    }
 };
 
 const addNewTextResource = async function ( ) {
 
+    if ( await getPermission( getCurrentWorkspace().workspaceId ) ){
+        ( debug ) ? console.log( "addNewTextResource() : Start" ) : null;
 
-    ( debug ) ? console.log( "addNewTextResource() : Start" ) : null;
+        // create a new resource
+        let resource = createNewResource();
 
-    // create a new resource
-    let resource = createNewResource();
+        // save the resource
+        await saveResource( resource );
 
-    // save the resource
-    await saveResource( resource );
+        // add the resource to the current topic
+        getCurrentActiveTopic().resources.push( resource );
 
-    // add the resource to the current topic
-    getCurrentActiveTopic().resources.push( resource );
-
-    // save the topic
-    await saveActiveTopic( );
+        // save the topic
+        await saveActiveTopic( );
 
 
-    ( debug ) ? console.log( "addNewTextResource() : Complete" ) : null;
+        ( debug ) ? console.log( "addNewTextResource() : Complete" ) : null;
+    }
 };
 
 function saveTextResource( resource, content ) {
